+++ src/main/java/org/folio/inventory/dataimport/handlers/matching/loaders/HoldingLoader.java
@@ -8,7 +8,6 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
-import io.vertx.core.Vertx;
 import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
 
@@ -25,12 +24,13 @@ import org.folio.rest.jaxrs.model.EntityType;
 public class HoldingLoader extends AbstractLoader<HoldingsRecord> {
 
   private static final String HOLDINGS_FIELD = "holdings";
+  private static final String INSTANCE_ID_FIELD = "instanceId";
+  private static final String INSTANCES_IDS_KEY = "INSTANCES_IDS";
 
   private Storage storage;
   private AbstractPreloader preloader;
 
-  public HoldingLoader(Storage storage, Vertx vertx, AbstractPreloader preloader) {
-    super(vertx);
+  public HoldingLoader(Storage storage, AbstractPreloader preloader) {
     this.storage = storage;
     this.preloader = preloader;
   }
@@ -57,6 +57,8 @@ public class HoldingLoader extends AbstractLoader<HoldingsRecord> {
     if (eventPayload.getContext() != null) {
       if (isNotEmpty(eventPayload.getContext().get(AbstractLoader.MULTI_MATCH_IDS))) {
         cqlSubMatch = getConditionByMultiMatchResult(eventPayload);
+      } else if (isNotEmpty(eventPayload.getContext().get(INSTANCES_IDS_KEY))) {
+        cqlSubMatch = getConditionByMultipleMarcBibMatchResult(eventPayload);
       } else if (isNotEmpty(eventPayload.getContext().get(EntityType.HOLDINGS.value()))) {
         JsonObject holdingAsJson = new JsonObject(eventPayload.getContext().get(EntityType.HOLDINGS.value()));
         if (holdingAsJson.getJsonObject(HOLDINGS_FIELD) != null) {
@@ -69,6 +71,10 @@ public class HoldingLoader extends AbstractLoader<HoldingsRecord> {
       }
     }
     return cqlSubMatch;
+  }
+
+  private String getConditionByMultipleMarcBibMatchResult(DataImportEventPayload eventPayload) {
+    return getConditionByMultipleValues(INSTANCE_ID_FIELD, eventPayload, INSTANCES_IDS_KEY);
   }
 
   @Override